aws_smithy_checksums

Module http

Checksum support for HTTP requests and responses.

Constants

CHECKSUM_ALGORITHMS_IN_PRIORITY_ORDER

When a response has to be checksum-verified, we have to check possible headers until we find the header with the precalculated checksum. Because a service may send back multiple headers, we have to check them in order based on how fast each checksum is to calculate.

CRC_32_C_HEADER_NAME

CRC_32_HEADER_NAME

CRC_64_NVME_HEADER_NAME

SHA_1_HEADER_NAME

SHA_256_HEADER_NAME

Traits

HttpChecksum

Checksum algorithms are use to validate the integrity of data. Structs that implement this trait can be used as checksum calculators. This trait requires Send + Sync because these checksums are often used in a threaded context.